Get a table of metadata for meteorological stations

Obtain a tibble containing information on all of the stations that have data.

get_station_metadata()

Value

Returns a tibble with 16 columns.

id

A character string that is a unique identifier for the weather station.

usaf

A character string identifying the fixed weather station from the USAF Master Station Catalog. USAF is an acronym for United States Air Force.

wban

A character string for the fixed weather station NCDC WBAN identifier. NCDC is an acronym for National Climatic Data Center. WBAN is an acronym for Weather Bureau, Air Force and Navy.

name

A character string with the station name.

country

A character string with the two character country code where the station is located. Not identical to country_code.

state

Character string of the two character abbreviation of a US state (when applicable).

icao

The ICAO identifier for the station.

lat

Latitude (degrees) rounded to three decimal places.

lon

Longitude (degrees) rounded to three decimal places.

elev

Numeric value for the elevation as measured in meters. The minimum value is -400 with a maximum of 8850. Elevation in feet can be approximated by elev * 3.28084

begin_date

The earliest date for which data are available.

end_date

The latest date for which data are available.

begin_year

The earliest year for which data are available.

end_year

The latest year for which data are available.

tz_name

The time zone name.

Examples

# Obtain a data frame with all available met stations
get_station_metadata()

#> # A tibble: 29,743 x 16
#>    id    usaf  wban  name  country state icao    lat   lon  elev begin_date
#>    <chr> <chr> <chr> <chr> <chr>   <chr> <chr> <dbl> <dbl> <dbl> <date>    
#>  1 0070… 0070… 99999 WXPO… <NA>    <NA>  <NA>    0     0   7018  2011-03-09
#>  2 0070… 0070… 99999 WXPO… AF      <NA>  <NA>    0     0   7026  2012-07-13
#>  3 0070… 0070… 99999 WXPO… AF      <NA>  <NA>    0     0   7070  2014-09-23
#>  4 0082… 0082… 99999 WXPO… <NA>    <NA>  <NA>    0     0      0  2005-01-01
#>  5 0082… 0082… 99999 WXPO… AF      <NA>  <NA>   32.9  65.6 1157. 2010-05-19
#>  6 0083… 0083… 99999 WXPO… AF      <NA>  <NA>    0     0   8318  2010-04-21
#>  7 0084… 0084… 99999 XM20  <NA>    <NA>  <NA>   NA    NA     NA  2016-02-17
#>  8 0084… 0084… 99999 XM18  <NA>    <NA>  <NA>   NA    NA     NA  2016-02-16
#>  9 0084… 0084… 99999 XM21  <NA>    <NA>  <NA>   NA    NA     NA  2016-02-17
#> 10 0084… 0084… 99999 XM24  <NA>    <NA>  <NA>   NA    NA     NA  2016-02-17
#> # … with 29,733 more rows, and 5 more variables: end_date <date>,
#> #   begin_year <int>, end_year <int>, tz_name <chr>, years <list>